+++ examples/favicon.ico
