+++ examples/fractal.py
