+++ examples/fractal.py.html
@@ -10,7 +10,7 @@
 </head>
 <body>
     <canvas id="canvas" width="512" height="512">no HTML5 support.</canvas>
-    <script src="./taichi.js"></script>
+    <script src="../taichi.js"></script>
     <script src="./fractal.py.js"></script>
     <script>
 let taichi = new Taichi(Module);
