+++ examples/tree_gravity.py
