+++ taichihub/__main__.py
@@ -1,5 +1,12 @@
-from . import app
+import socket
+import sys
 
 
 if __name__ == '__main__':
-    app.run(host='winner', port=80, debug=True)
+    if sys.argv[1] == 'compile':
+        from .compiler import do_compile
+        do_compile(sys.argv[2])
+
+    elif sys.argv[1] == 'server':
+        from .flaskapp import app
+        app.run(host=socket.gethostname(), port=int(sys.argv[2]), debug=True)
